Guests can't post to forums any longer, removed memberdata cache, members are now soft-deleted (removed old deletedmembers file / reserved names), bug fixes, ...

+++ install/package/install/db/abos.sql
@@ -1,6 +1,6 @@
 CREATE TABLE `{:=DBPREFIX=:}abos` (
   `id` bigint(12) unsigned NOT NULL auto_increment,
-  `mid` mediumint(7) unsigned NOT NULL,
+  `mid` int(7) unsigned NOT NULL,
   `tid` int(10) unsigned NOT NULL,
   `type` enum('','d','w','f') NOT NULL DEFAULT '',
   PRIMARY KEY  (`id`),
