+++ install/package/install/db/votes.sql
@@ -1,6 +1,6 @@
 CREATE TABLE `{:=DBPREFIX=:}votes` (
   `id` bigint(19) unsigned NOT NULL auto_increment,
-  `mid` mediumint(7) unsigned NOT NULL,
+  `mid` int(7) unsigned NOT NULL,
   `aid` mediumint(7) unsigned NOT NULL,
   PRIMARY KEY  (`id`),
   KEY `aid` (`aid`)
